Pretty much finished intro animations
all that is left for the intro is that there is a inconsistant bug that makes the intro text pop back up after the animation is done. i am 95% sure its just timeing that i have to fix but i need to go to sleep for work. so i leave this to future luis (as well as all the other shit)