Final strech?? hopefully
finshed the animation for the projects section. found out you cant do an animation with a height that either begins or ends with :auto. what i belive is last is expanding the background image of the project divs to cover everything, adding content to the project sections. the pplan is to give it a black background with low opacity at the bottom of the page. and lastly make the popup works